Allow method:get to be omitted.
Previously, omitting the method would lead to a bad HTTP with an empty
method being sent, resulting in a status 400.

This change also simplifies the MediaWikiActionApiEditTest example.
Besides omitting the method:get where possible, it now also makes
use of structured comparison of json responses instead of relying on
regular expressions.

For convenience, this change introduces a $default parameter
in Instructions::getLowerCase().

As an aside, this renames Instructions::getArray() to
Instructions::asArray() for consistency with other getts and with
hasArray().