Refine output format and behavior.
This changes when and where which information is written to output.
In particular, it causes the suite names to be shown regardless of
whether any errors occur.

This also changes the testing behavior so that a failure in the setup
will abort the suite, and a failure in a test will abort that test but
not the suite. A single assertion can however still produce multiple
error messages.